+++ src/app/components/features/recipe/RecipeList/RecipeListVertical.tsx
@@ -21,16 +21,3 @@ export const RecipeListVertical: FC<Props> = ({ recipeList }) => {
     </ul>
   );
 };
-
-export const RecipeListHorizontal: FC<Props> = ({ recipeList }) => {
-  const recipe = recipeList.length > 0 ? recipeList : dummyRecipeList;
-  return (
-    <ul className="hidden-scrollbar flex gap-x-4 overflow-x-scroll">
-      {recipe.map((recipe) => (
-        <li key={recipe.id} className="w-40 flex-shrink-0">
-          <RecipeListItem recipe={recipe} />
-        </li>
-      ))}
-    </ul>
-  );
-};
